Create the interface IOpenApiReader.
Split OpenApiParser into OpenApiStreamReader and OpenApiStringReader. Note the suffix change from Parser to Reader to make clear that this is the "first-level" reader, i.e. the public interface.

The OpenApiV2Reader and OpenApiV3Reader are technically NOT readers at this point. They build the document after being called by the reader, so I'll change their names to Builders.

It's still strange we are overloading ParsingContext as both context and logging (i.e. ParseErrors). I'll fix that separately since that will be a bigger change.