Rename ILog to IDiagnostic. Take out OpenApiDocument and Errors out of ParsingContext and pass around as a separate object.
Remove the requirement that IDiagnostic must contain IList of TError. It is too restrictive and may not apply for other readers.

The semantics for ParsingContext is now only things needed DURING the reading process.

The document from the reading process is in the return object, and all the information needed AFTER the reading process now goes into IDiagnostic (OpenApiDiagnostic for this reader).