Fixes #49
Fix logic for WriteAsVX for these objects:
- Document (still not complete)
- Server
- OAuthFlow
- SecurityScheme
- SecurityRequirement

Add Unit tests for
- Server
- SecurityScheme
- SecurityRequirement

Also, resolves #19. I am using FluentAssertions in the unit test now. The main reason is that I want a richer error message that tells us where the two strings are different. The built-in Assert from xUnit is not as good. This should help us write unit test faster.

Future tests can use FluentAssertions, and we can revise old test on an incremental basis.

If you strongly object FluentAssertions, we should discuss before we make more investments on this.