Re-enable the default rule count test case
The problem was this:

ValidateCustomExtension test calls ValidationRuleSet.DefaultRuleSet and makes changes to it directly. Since the DefaultRuleSet is cached, this means the private _defaultRuleSet is actually modified.

The reason we are not seeing this repro locally might be that the order of test running locally and the order on AppVeyor are different. We see the issue only if the "Count" test is run after the custom extension test.