Serialize explode value properly on parameters with style "form"
When a parameter has the style set to "form" and explode set to false,
the v3 serializer will not write out the value for explode. Swagger UI
then defaults the explode value to true, since it is not specified and
that is the correct default value when the style is set to "form". The
OpenAPI spec for this is right around here:

https://github.com/OAI/OpenAPI-Specification/blob/master/versions/3.0.0.md#fixed-fields-10